+++ tests/cnp_node_new_free.c
@@ -3,7 +3,7 @@
 #include <assert.h>
 
 
-size_t cnp_len = 1000;
+int cnp_len = 1000;
 copy_num max_copy_num;
 double **neighbor_probs;
 double **mutation_probs;
