+++ tests/phylogeny_optimize.c
@@ -4,7 +4,7 @@
 #include <string.h>
 
 
-size_t cnp_len = 5;
+int cnp_len = 5;
 copy_num max_copy_num = 2;
 double **neighbor_probs;
 double **mutation_probs;
